// src/app/main/interfaces/bonofecha.interface.ts
export interface BonoFecha {
    fechaProgramada?: string;
    bono?: number;
    bonoIndexado?: number;
    cupon?: number;
    cuota?: number;
    amortizacion?: number;
    prima?: number;
    escudo?: number;
    flujoEmisor?: number;
    flujoEmisorEscudo?: number;
    flujoBonista?: number;
    flujoActivo?: number;
    faPlazo?: number;
    factorConvexidad?: number;
}