// src/app/main/interfaces/resultado.interface.ts
export interface Resultado {
    id?: number;
    frecuenciaCupon?: number;
    diasCapitalizacion: number;
    numeroPeriodosAno: number;
    numeroTotalPeriodos: number;
    tasaEfectivaAnual: number;
    tasaEfectivaPeriodo: number;
    cok: number;
    costesInicialesEmisor: number;
    costesInicialesBonista: number;
    bonoId: number
}